+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Fendo
   - Lines of Action
   - Manalath
+  - Martian Chess
 
 ## [0.4.0] - 2021-11-15